refactor(pgn): consolidate PGN chunk handling into a dedicated module
- Removed the `bench-pgn-chunks-iterator.ts` and `pgn-chunks-generator.ts` files to streamline the codebase.
- Introduced a new `pgn-chunks.ts` module that centralizes PGN chunk reading and processing logic, enhancing maintainability and clarity.
- Updated references across the codebase to utilize the new `pgn-chunks` module for chunk handling, ensuring consistent functionality.
